% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_credit_balance}
\alias{get_credit_balance}
\title{Get AI gateway credit balance}
\usage{
get_credit_balance(client)
}
\arguments{
\item{client}{A dickless client}
}
\value{
List with balanceCents, lifetimePurchasedCents, etc.
}
\description{
Get AI gateway credit balance
}
\examples{
\dontrun{
client <- dickless_client(api_key = Sys.getenv("DICKLESS_API_KEY"))
balance <- get_credit_balance(client)
}
}
