% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_credit_transactions}
\alias{get_credit_transactions}
\title{Get AI gateway credit transactions}
\usage{
get_credit_transactions(client)
}
\arguments{
\item{client}{A dickless client}
}
\value{
List of transaction records
}
\description{
Get AI gateway credit transactions
}
\examples{
\dontrun{
client <- dickless_client(api_key = Sys.getenv("DICKLESS_API_KEY"))
transactions <- get_credit_transactions(client)
}
}
