% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_short_url_stats}
\alias{get_short_url_stats}
\title{Get short URL click statistics}
\usage{
get_short_url_stats(client, code)
}
\arguments{
\item{client}{A dickless client}

\item{code}{Short URL code}
}
\value{
List with code, targetUrl, clicks, createdAt
}
\description{
Get short URL click statistics
}
\examples{
\dontrun{
client <- dickless_client(api_key = Sys.getenv("DICKLESS_API_KEY"))
stats <- get_short_url_stats(client, "abc123")
}
}
